fix: remove some code duplication for miners
