+++ finetuner/tuner/paddle/miner.py
@@ -3,7 +3,8 @@ from typing import Tuple
 
 import paddle
 
-from ..base import BaseClassMiner, BaseSessionMiner
+from ..miner import get_session_pairs, get_session_triplets
+from ..miner.base import BaseClassMiner, BaseSessionMiner
 
 
 def _empty_tensor(dtype: str = 'int64') -> paddle.Tensor:
@@ -116,20 +117,7 @@ class SiameseSessionMiner(BaseSessionMiner[paddle.Tensor]):
         assert len(distances) == len(labels[0]) == len(labels[1])
 
         sessions, match_types = [x.tolist() for x in labels]
-        ind_one, ind_two, labels_ret = [], [], []
-
-        labels_index = [
-            (sess, match_type, index)
-            for index, (sess, match_type) in enumerate(zip(sessions, match_types))
-        ]
-        sorted_labels_with_index = sorted(labels_index, key=lambda x: x[0])
-
-        for _, group in groupby(sorted_labels_with_index, lambda x: x[0]):
-            for left, right in combinations(group, 2):  # (session_id, label, ind)
-                if left[1] != -1 or right[1] != -1:
-                    ind_one.append(left[2])
-                    ind_two.append(right[2])
-                    labels_ret.append(0 if min(left[1], right[1]) == -1 else 1)
+        ind_one, ind_two, labels_ret = get_session_pairs(sessions, match_types)
 
         return (
             paddle.to_tensor(ind_one, place=distances.place),
@@ -157,28 +145,7 @@ class TripletSessionMiner(BaseSessionMiner[paddle.Tensor]):
         assert len(distances) == len(labels[0]) == len(labels[1])
 
         sessions, match_types = [x.tolist() for x in labels]
-        anchor_ind, pos_ind, neg_ind = [], [], []
-
-        labels_index = [
-            (sess, match_type, index)
-            for index, (sess, match_type) in enumerate(zip(sessions, match_types))
-        ]
-        sorted_labels_with_index = sorted(labels_index, key=lambda x: x[0])
-
-        for _, group in groupby(sorted_labels_with_index, lambda x: x[0]):
-            anchor_pos_session_indices = []
-            neg_session_indices = []
-
-            for _, session_label, session_index in group:
-                if session_label >= 0:
-                    anchor_pos_session_indices.append(session_index)
-                else:
-                    neg_session_indices.append(session_index)
-
-            for anchor, pos in permutations(anchor_pos_session_indices, 2):
-                anchor_ind += [anchor] * len(neg_session_indices)
-                pos_ind += [pos] * len(neg_session_indices)
-                neg_ind += neg_session_indices
+        anchor_ind, pos_ind, neg_ind = get_session_triplets(sessions, match_types)
 
         return (
             paddle.to_tensor(anchor_ind, place=distances.place),
